--- a/path.go
+++ b/path.go
@@ -15,91 +15,141 @@ import (
 
 // routeParser  holds the path segments and param names
 type routeParser struct {
-	segs   []paramSeg
+	segs   []routeSegment
 	params []string
 }
 
 // paramsSeg holds the segment metadata
-type paramSeg struct {
+type routeSegment struct {
 	Param      string
 	Const      string
 	IsParam    bool
+	IsWildcard bool
 	IsOptional bool
 	IsLast     bool
-	EndChar    byte
+	EndChar    byte // TODO: remove ?
 }
 
-// list of possible parameter and segment delimiter
-// slash has a special role, unlike the other parameters it must not be interpreted as a parameter
-// TODO '(' ')' delimiters for regex patterns
-var routeDelimiter = []byte{'/', '-', '.'}
+const (
+	wildcardParam    byte = '*'
+	optionalParam    byte = '?'
+	slashDelimiter   byte = '/'
+	paramStarterChar byte = ':'
+)
 
-const wildcardParam string = "*"
+var (
+	// list of possible parameter and segment delimiter
+	// slash has a special role, unlike the other parameters it must not be interpreted as a parameter
+	// TODO '(' ')' delimiters for regex patterns
+	routeDelimiter = []byte{slashDelimiter, '-', '.'}
+	// list of chars for the parameter recognising
+	parameterStartChars = []byte{wildcardParam, paramStarterChar}
+	// list of chars to find the end of a parameter
+	parameterDelimiterChars = append([]byte{paramStarterChar}, routeDelimiter...)
+	// list of chars to find the end of a parameter
+	parameterEndChars = append([]byte{optionalParam}, parameterDelimiterChars...)
+)
 
 // parseRoute analyzes the route and divides it into segments for constant areas and parameters,
 // this information is needed later when assigning the requests to the declared routes
-func parseRoute(pattern string) (p routeParser) {
-	var out []paramSeg
+func parseRoute(pattern string) routeParser {
+	var segList []routeSegment
 	var params []string
 
-	part, delimiterPos := "", 0
-	for len(pattern) > 0 && delimiterPos != -1 {
-		delimiterPos = findNextRouteSegmentEnd(pattern)
-		if delimiterPos != -1 {
-			part = pattern[:delimiterPos]
+	part := ""
+	for len(pattern) > 0 {
+		nextParamPosition := findNextParamPosition(pattern)
+		// handle the parameter part
+		if nextParamPosition == 0 {
+			processedPart, seg := analyseParameterPart(pattern)
+			params, segList, part = append(params, seg.Param), append(segList, seg), processedPart
 		} else {
-			part = pattern
+			processedPart, seg := analyseConstantPart(pattern, nextParamPosition)
+			segList, part = append(segList, seg), processedPart
 		}
 
-		partLen, lastSeg := len(part), len(out)-1
-		if partLen == 0 { // skip empty parts
-			if len(pattern) > 0 {
-				// remove first char
-				pattern = pattern[1:]
-			}
-			continue
+		// reduce the pattern by the processed parts
+		if len(part) == len(pattern) {
+			break
 		}
-		// is parameter ?
-		if part[0] == '*' || part[0] == ':' {
-			out = append(out, paramSeg{
-				Param:      utils.GetTrimmedParam(part),
-				IsParam:    true,
-				IsOptional: part == wildcardParam || part[partLen-1] == '?',
-			})
-			lastSeg = len(out) - 1
-			params = append(params, out[lastSeg].Param)
-			// combine const segments
-		} else if lastSeg >= 0 && !out[lastSeg].IsParam {
-			out[lastSeg].Const += string(out[lastSeg].EndChar) + part
-			// create new const segment
-		} else {
-			out = append(out, paramSeg{
-				Const: part,
-			})
-			lastSeg = len(out) - 1
-		}
-
-		if delimiterPos != -1 && len(pattern) >= delimiterPos+1 {
-			out[lastSeg].EndChar = pattern[delimiterPos]
-			pattern = pattern[delimiterPos+1:]
-		} else {
-			// last default char
-			out[lastSeg].EndChar = '/'
-		}
+		pattern = pattern[len(part):]
 	}
-	if len(out) > 0 {
-		out[len(out)-1].IsLast = true
+	// mark last segment
+	if len(segList) > 0 {
+		segList[len(segList)-1].IsLast = true
 	}
 
-	p = routeParser{segs: out, params: params}
-	return
+	return routeParser{segs: segList, params: params}
 }
 
-// findNextRouteSegmentEnd searches in the route for the next end position for a segment
-func findNextRouteSegmentEnd(search string) int {
+// findNextParamPosition search for the next possible parameter start position
+func findNextParamPosition(pattern string) int {
+	nextParamPosition := findNextCharsetPosition(pattern, parameterStartChars)
+	if nextParamPosition != -1 && len(pattern) > nextParamPosition && pattern[nextParamPosition] != wildcardParam {
+		// search for parameter characters for the found parameter start,
+		// if there are more, move the parameter start to the last parameter char
+		for found := findNextCharsetPosition(pattern[nextParamPosition+1:], parameterStartChars); found == 0; {
+			nextParamPosition++
+			if len(pattern) > nextParamPosition {
+				break
+			}
+		}
+	}
+
+	return nextParamPosition
+}
+
+// analyseConstantPart find the end of the constant part and create the route segment
+func analyseConstantPart(pattern string, nextParamPosition int) (string, routeSegment) {
+	// handle the constant part
+	processedPart := pattern
+	if nextParamPosition != -1 {
+		// remove the constant part until the parameter
+		processedPart = pattern[:nextParamPosition]
+	}
+	return processedPart, routeSegment{
+		Const: processedPart,
+	}
+}
+
+// analyseParameterPart find the parameter end and create the route segment
+func analyseParameterPart(pattern string) (string, routeSegment) {
+	isWildCard := pattern[0] == wildcardParam
+	parameterEndPosition := findNextCharsetPosition(pattern[1:], parameterEndChars)
+	// handle wildcard end
+	if isWildCard {
+		parameterEndPosition = 0
+	} else if parameterEndPosition == -1 {
+		parameterEndPosition = len(pattern) - 1
+	} else if false == isInCharset(pattern[parameterEndPosition+1], parameterDelimiterChars) {
+		parameterEndPosition = parameterEndPosition + 1
+	}
+	// cut params part
+	processedPart := pattern[0 : parameterEndPosition+1]
+
+	return processedPart, routeSegment{
+		Param:      utils.GetTrimmedParam(processedPart),
+		IsParam:    true,
+		IsOptional: isWildCard || pattern[parameterEndPosition] == optionalParam,
+		IsWildcard: isWildCard,
+	}
+}
+
+// isInCharset check is the given character in the charset list
+func isInCharset(searchChar byte, charset []byte) bool {
+	for _, char := range charset {
+		if char == searchChar {
+			return true
+		}
+	}
+	return false
+}
+
+// findNextCharsetPosition search the next char position from the charset
+func findNextCharsetPosition(search string, charset []byte) int {
 	nextPosition := -1
-	for _, delimiter := range routeDelimiter {
-		if pos := strings.IndexByte(search, delimiter); pos != -1 && (pos < nextPosition || nextPosition == -1) {
+	for _, char := range charset {
+		if pos := strings.IndexByte(search, char); pos != -1 && (pos < nextPosition || nextPosition == -1) {
 			nextPosition = pos
 		}
 	}
@@ -111,33 +161,18 @@ func findNextRouteSegmentEnd(search string) int {
 func (p *routeParser) getMatch(s string, partialCheck bool) ([][2]int, bool) {
 	lenKeys := len(p.params)
 	paramsPositions := getAllocFreeParamsPos(lenKeys)
-	var i, j, paramsIterator, partLen, paramStart int
-	if len(s) > 0 {
-		s = s[1:]
-		paramStart++
-	}
+	var i, paramsIterator, partLen, paramStart int
+	//if len(s) > 0 {
+	//	s = s[1:]
+	//	paramStart++
+	//}
 	for index, segment := range p.segs {
 		partLen = len(s)
 		// check parameter
 		if segment.IsParam {
 			// determine parameter length
-			if segment.Param == wildcardParam {
-				if segment.IsLast {
-					i = partLen
-				} else {
-					i = findWildcardParamLen(s, p.segs, index)
-				}
-			} else {
-				i = strings.IndexByte(s, segment.EndChar)
-			}
-			if i == -1 {
-				i = partLen
-			}
-
+			i = findParamLen(s, p.segs, index)
 			if !segment.IsOptional && i == 0 {
-				return nil, false
-				// special case for not slash end character
-			} else if i > 0 && partLen >= i && segment.EndChar != '/' && s[i-1] == '/' {
 				return nil, false
 			}
 
@@ -145,21 +180,28 @@ func (p *routeParser) getMatch(s string, partialCheck bool) ([][2]int, bool) {
 			paramsIterator++
 		} else {
 			// check const segment
+			optionalPart := false
 			i = len(segment.Const)
-			if partLen < i || (i == 0 && partLen > 0) || s[:i] != segment.Const || (partLen > i && s[i] != segment.EndChar) {
+			if i > 0 && partLen == i-1 && segment.Const[i-1] == slashDelimiter && s[:i-1] == segment.Const[:i-1] {
+				if segment.IsLast || p.segs[index+1].IsOptional {
+					i--
+					optionalPart = true
+				}
+			}
+
+			if optionalPart == false && (partLen < i || (i == 0 && partLen > 0) || s[:i] != segment.Const) {
 				return nil, false
 			}
 		}
 
 		// reduce founded part from the string
 		if partLen > 0 {
-			j = i + 1
-			if segment.IsLast || partLen < j {
-				j = i
+			if partLen < i {
+				i = partLen
 			}
-			paramStart += j
+			paramStart += i
 
-			s = s[j:]
+			s = s[i:]
 		}
 	}
 	if len(s) != 0 && !partialCheck {
@@ -184,29 +226,42 @@ func (p *routeParser) paramsForPos(path string, paramsPositions [][2]int) []stri
 	return params
 }
 
-// findWildcardParamLen for the expressjs wildcard behavior (right to left greedy)
+// findParamLen for the expressjs wildcard behavior (right to left greedy)
 // look at the other segments and take what is left for the wildcard from right to left
-func findWildcardParamLen(s string, segments []paramSeg, currIndex int) int {
+func findParamLen(s string, segments []routeSegment, currIndex int) int {
+	if segments[currIndex].IsLast {
+		if segments[currIndex].IsWildcard {
+			return len(s)
+		}
+		if i := strings.IndexByte(s, slashDelimiter); i != -1 {
+			return i
+		}
+
+		return len(s)
+	}
 	// "/api/*/:param" - "/api/joker/batman/robin/1" -> "joker/batman/robin", "1"
 	// "/api/*/:param" - "/api/joker/batman"         -> "joker", "batman"
 	// "/api/*/:param" - "/api/joker-batman-robin/1" -> "joker-batman-robin", "1"
-	endChar := segments[currIndex].EndChar
-	neededEndChars := 0
-	// count the needed chars for the other segments
-	for i := currIndex + 1; i < len(segments); i++ {
-		if segments[i].EndChar == endChar {
-			neededEndChars++
+	nextSeg := segments[currIndex+1]
+	// check next segment
+	if nextSeg.IsParam {
+		if segments[currIndex].IsWildcard || nextSeg.IsWildcard {
+			// greedy logic
+			for i := currIndex + 1; i < len(segments); i++ {
+				if false == segments[i].IsParam {
+					nextSeg = segments[i]
+					break
+				}
+			}
+		} else if len(s) > 0 {
+			// in case the next parameter or the current parameter is not a wildcard its not greedy, we only want one character
+			return 1
 		}
 	}
-	// remove the part the other segments still need
-	for {
-		pos := strings.LastIndexByte(s, endChar)
-		if pos != -1 {
-			s = s[:pos]
-		}
-		neededEndChars--
-		if neededEndChars <= 0 || pos == -1 {
-			break
+	// get the length to the next constant part
+	if false == nextSeg.IsParam {
+		if constPosition := strings.Index(s, nextSeg.Const); constPosition != -1 {
+			return constPosition
 		}
 	}
 
@@ -221,6 +276,7 @@ var paramsDummy, paramsPosDummy = make([]string, 100000), make([][2]int, 100000)
 // to assign a separate range to each request
 var startParamList, startParamPosList uint32 = 0, 0
 
+// TODO: replace it with bytebufferpool and release the parameter buffers in ctx release function
 // getAllocFreeParamsPos fetches a slice area from the predefined slice, which is currently not in use
 func getAllocFreeParamsPos(allocLen int) [][2]int {
 	size := uint32(allocLen)